Prototype test of what can be moved into pgrustxn.
	TODO add jason to author

-Alex
