+++ pgrustxn/Cargo.toml
@@ -1,8 +1,7 @@
 [package]
 name = "pgrustxn"
 version = "0.0.1"
-authors = ["Alex Newman <posi@planet.com>",
-           "Jason Dusek <jason.dusek@gmail.com>"]
+authors = ["Alex Newman <posi@planet.com>" ]
 
 [dev-dependencies]
 rustfmt = "*"
