+++ pgrustxn/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod pg94;
 pub mod pg95;
-
+pub mod texttools;
 pub mod decoder;
